Force moves to always use 16-bit immediate moves.
The most predictable way to load a 32-bit immediate value into a
register is to combine the 16-bit MOV with 16-bit MOVT (when there are
any top bits to set).

Just issuing MOV doesn't always work, because as tries to get clever
about which exact MOV variation it picks.

So I issue the exact 32-bit word I want, basically writing the crude
beginnings of an assembler.

This is still not perfectly efficient.  Some values which we issue two
instructions for may be encodable using the 12-bit immediate, 4-bit
rotation variant.  But this version is reliable and way faster than the
previous version using addition.