Working on stack discipline.
Had been cargo-culting the return address on the stack.  It's really in
LR, which means we can begin si at 0 (not -4).

Also, formalize that when we *do* save the link register, we just
decrement (and later increment) sp.  I believe it'll simplify things
down the road (unsure if adding these SP modifications slows things
down).