Closures now work.
I used an extremely inefficient convention for dealing with the closure
pointer.  It gets placed on the stack *and* loaded into a register.  We
untag the closure object, then store it at a well-defined place on the
stack.  Procedures load it into the closure register, and after a
subroutine returns, we restore it from that spot on the stack.