Rip out the old 'datum' forms and trying to place constants in BSS.
It's the right way to do it eventually, but switching to this approach
is much simpler as a first implementation.  Re-doing this also made it
easy to untangle the mess that had accumulated with the datum form, the
insanity of collect-code, and yet more cruft that never belonged in
parse-and-rename.